requirements.txt
asn1crypto==1.5.1
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==36.0.2
Deprecated==1.2.13
dnspython==2.2.1
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
gcloud==0.18.3
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
MarkupSafe==2.1.1
oauth2client==4.1.3
gunicorn
oscrypto==1.3.0
outcome==1.2.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyJWT==2.4.0
pymongo==4.2.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
Pyrebase4==4.5.0
PySocks==1.7.1
python-dotenv==0.21.0
python-jwt==3.3.4
pytube==12.1.0
pytz==2022.1
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.9
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==2.7.9
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
webdriver-manager==3.8.3
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
wsproto==1.2.0
